spread/client, spread/project: workaround bash 4.3 subshell exit status
Parts of the stage script (prepare, restore..) are assembled into one script
using subshells. REBOOT and ERROR helpers make the shell exit with special
status code 213 which also counts as an error exit status. Normally, we expect
the error exit status to bubble up and be observed by the SSH session runner.

However, due to bug in bash 4.3, this does not happen on systems using that
particular version. The set -e switch causes upper subshell to exit with status
1, instead of forwarding the exit status of the inner shell, provided the inner
one is not the last.

Thus an example script like this:

  set -eu
  set -x
  ( # stage

  (
    # snippet, eg. restore
    REBOOT
  )

  (
   # snippet, eg. restore-each
  )
  ) < /dev/null

Exits with status 1, instead of 213. This can be reproduced with minimal setup:

restore-each: |
    echo foo

restore: |
    echo doing reboot
    REBOOT

On a system with bash 4.3 (eg. Ubuntu 16.04), REBOOT is not executed during
restore.

Signed-off-by: Maciek Borzecki <maciek.borzecki@gmail.com>