Comment the config in case there vars are already set
Comemnt the lines in case there is a file with the PermitRootLogin or
PasswordAuthentication. As the command could fail when there is not
files in sshd_config.d dir it is added a || true to make sure the script
is not going to fail because of that.

Also it is being updated the name of the config file